Added working two click share buttons. Fixes #163. Code clean up.
