+++ assets/css/route.css
@@ -1,3 +1,6 @@
+a.disabled { 
+  text-decoration:none; color:#bbb; 
+}
 
 @media (min-width: 767px){
   .graph-selection {
@@ -140,4 +143,97 @@ img.olTileImage {
 
 #statistics > li {
   padding-top: 10px;
+}
+
+.onoffswitch {
+  position: absolute;
+  right: 10px;
+  bottom: 45px; 
+  width: 106px;
+  -webkit-user-select:none; 
+  -moz-user-select:none; 
+  -ms-user-select: none;
+}
+.onoffswitch-checkbox {
+  display: none;
+}
+.onoffswitch-label {
+  display: block; 
+  overflow: hidden; 
+  cursor: pointer;
+  border: 0px solid #999999; 
+  border-radius: 0px;
+}
+.onoffswitch-inner {
+  width: 200%; 
+  margin-left: -100%;
+  -moz-transition: margin 0.3s ease-in 0s; 
+  -webkit-transition: margin 0.3s ease-in 0s;
+  -o-transition: margin 0.3s ease-in 0s; 
+  transition: margin 0.3s ease-in 0s;
+}
+.onoffswitch-inner > div {
+  float: left; 
+  position: relative; 
+  width: 50%; 
+  height: 30px; 
+  padding: 0; 
+  line-height: 30px;
+  font-size: 14px; 
+  color: white; 
+  font-family: Trebuchet, Arial, sans-serif; 
+  font-weight: bold;
+  -moz-box-sizing: border-box; 
+  -webkit-box-sizing: border-box; 
+  box-sizing: border-box;
+}
+.onoffswitch-inner .onoffswitch-active {
+  padding-left: 15px;
+  background-color: #C2C2C2; 
+  color: #FFFFFF;
+}
+.onoffswitch-inner .onoffswitch-inactive {
+  padding-right: 15px;
+  background-color: #C2C2C2; 
+  color: #FFFFFF;
+  text-align: right;
+}
+.onoffswitch-switch {
+  width: 40px; 
+  margin: 0px; 
+  text-align: center;
+  border: 0px solid #999999;
+  border-radius: 0px;
+  position: absolute; 
+  top: 0; 
+  bottom: 0;
+}
+.onoffswitch-active .onoffswitch-switch {
+  background: #27A1CA; 
+  left: 0;
+}
+.onoffswitch-inactive .onoffswitch-switch {
+  background: #A1A1A1; 
+  right: 0;
+}
+.onoffswitch-active .onoffswitch-switch:before {
+  content: " "; 
+  position: absolute; 
+  top: 0; 
+  left: 40px;
+  border-style: solid; 
+  border-color: #27A1CA transparent transparent #27A1CA; 
+  border-width: 15px 10px;
+}
+.onoffswitch-inactive .onoffswitch-switch:before {
+  content: " "; 
+  position: absolute; 
+  top: 0; 
+  right: 40px;
+  border-style: solid; 
+  border-color: transparent #A1A1A1 #A1A1A1 transparent; 
+  border-width: 15px 10px;
+}
+.onoffswitch-checkbox:checked + .onoffswitch-label .onoffswitch-inner {
+  margin-left: 0;
 }
