+++ assets/js/about.js.php
@@ -1,3 +1,4 @@
+<script type="text/javascript">
 (function ($) {
 
 	$(function(){
@@ -27,3 +28,4 @@
 	});
 
 })(window.jQuery);
+</script>
