--- assets/js/dashboard.js
@@ -1,5 +1,0 @@
-$(function(){
-       $('body').tooltip({
-		    selector: '[rel=tooltip]'
-		});
-});
