--- assets/js/editprofile.js
@@ -1,2 +1,0 @@
-//enable form validation
-$(function () { $("input,select,textarea").not("[type=submit]").jqBootstrapValidation(); } );
