+++ assets/js/editprofile.js.php
@@ -1,0 +1,4 @@
+<script type="text/javascript">
+//enable form validation
+$(function () { $("input,select,textarea").not("[type=submit]").jqBootstrapValidation(); } );
+</script>
