+++ assets/js/registration.js.php
@@ -1,0 +1,2 @@
+<script type="text/javascript">
+</script>
